Fix uint8 overflow in drawMonoPackedPixels byte index
byteIdx was uint8_t (max 255) but needs to reach 359 for 2880 pixels.
Overflowed at pixel 2048 (row 28), wrapping back to buffer start and
corrupting the bottom of the image.

+++ arduino/ltp_esp32c3_oled/oled_driver.h
@@ -102,7 +102,7 @@ public:
             uint8_t y = i / OLED_WIDTH;
 
             // MSB-first: bit 7 of byte 0 = pixel 0
-            uint8_t byteIdx = i >> 3;
+            uint16_t byteIdx = i >> 3;
             uint8_t bitIdx = 7 - (i & 7);
             bool on = (packedBuffer[byteIdx] >> bitIdx) & 1;
